add services Opportunity::Filter
