+++ composer.json
@@ -23,7 +23,8 @@
             "database/factories"
         ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "App\\Services\\": "app/Services/"
         }
     },
     "autoload-dev": {
