+++ resources/views/opportunity/filter.blade.php
@@ -1,0 +1,1 @@
+@include('opportunity._filter_result')
